optimal updateAssignment feature

+++ src/main/java/com/laptrinhjavaweb/repository/CustomerRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 public interface CustomerRepository extends CustomerRepositoryCustom, JpaRepository<CustomerEntity, Long> {
     void deleteByIdIn(List<Long> ids);
 
-    Long countByIdIn(List<Long> ids);
+    int countByIdIn(List<Long> ids);
 }
